@@ -11,18 +11,20 @@
 
 package org.lfenergy.letscoordinate.backend.config;
 
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
+import org.lfenergy.letscoordinate.backend.enums.ChangeJsonDataFromWhichEnum;
 import org.lfenergy.letscoordinate.backend.enums.UnknownEicCodesProcessEnum;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 @ConfigurationProperties(prefix = "letsco")
 @Getter
 @Setter
 public class LetscoProperties {
+
     private InputFile inputFile;
 
     @Getter
@@ -34,10 +36,41 @@ public class LetscoProperties {
         @Getter
         @Setter
         public static class Validation {
+            @Getter(AccessLevel.NONE)
+            private List<String> ignoreProcesses = new ArrayList<>();
+            @Getter(AccessLevel.NONE)
+            private Map<String, ChangeSource> changeSource;
+            @Getter(AccessLevel.NONE)
+            private Map<String, String> changeMessageTypeName = new HashMap<>();
+            private boolean businessDayFromOptional;
             private boolean acceptPropertiesIgnoreCase;
             private boolean failOnUnknownProperties;
             private UnknownEicCodesProcessEnum unknownEicCodesProcess = UnknownEicCodesProcessEnum.EXCEPTION;
-            private List<String> allowedEicCodes = new ArrayList<>();
+            @Getter(AccessLevel.NONE)
+            private List<String> allowedEicCodes;
+
+            public Optional<Map<String, ChangeSource>> getChangeSource() {
+                return Optional.ofNullable(changeSource);
+            }
+
+            public Optional<List<String>> getIgnoreProcesses() {
+                return Optional.ofNullable(ignoreProcesses);
+            }
+
+            public Optional<Map<String, String>> getChangeMessageTypeName() {
+                return Optional.ofNullable(changeMessageTypeName);
+            }
+
+            public Optional<List<String>> getAllowedEicCodes() {
+                return Optional.ofNullable(allowedEicCodes);
+            }
+        }
+
+        @Getter
+        @Setter
+        public static class ChangeSource {
+            private ChangeJsonDataFromWhichEnum fromWhichLevel;
+            private String changingField;
         }
     }
 }
