@@ -33,9 +33,9 @@ public final class StringUtil {
     public static final String PROCESS_INFORMATION = "ProcessInformation";
     public static final String MESSAGE_VALIDATED = "DfgMessageValidated";
 
-    public static final String POSITIVE_ACK = "Positive ACK";
-    public static final String POSITIVE_ACK_WITH_WARNINGS = "Positive ACK with warnings";
-    public static final String NEGATIVE_ACK = "Negative ACK";
+    public static final String POSITIVE_ACK = "Positive";
+    public static final String POSITIVE_ACK_WITH_WARNINGS = "Positive with warnings";
+    public static final String NEGATIVE_ACK = "Negative";
 
     /**
      * <p>This function allows to split an input string according to a separator after removing all spaces from it
